Refactor to avoid DRY violations
