+++ src/DashboardUI/src/components/QuickSearchToolbar.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import {
+  GridToolbarContainer,
+  GridToolbarQuickFilter,
+  GridToolbarFilterButton,
+  GridToolbarExport,
+} from '@mui/x-data-grid';
+
+function QuickSearchToolbar() {
+  return (
+    <GridToolbarContainer>
+      <GridToolbarQuickFilter />
+      <GridToolbarFilterButton />
+      <GridToolbarExport />
+    </GridToolbarContainer>
+  );
+}
+
+export default QuickSearchToolbar;
